+++ src/datasource/components/QueryEditor.tsx
@@ -1,6 +1,17 @@
 import defaults from 'lodash/defaults';
 import React, { PureComponent } from 'react';
-import { Alert, Icon, InlineField, InlineFieldRow, LinkButton, MultiSelect, Select, Input } from '@grafana/ui';
+import {
+  Alert,
+  FieldValidationMessage,
+  Icon,
+  InlineField,
+  InlineFieldRow,
+  InlineSwitch,
+  Input,
+  LinkButton,
+  MultiSelect,
+  Select,
+} from '@grafana/ui';
 import { QueryEditorProps, SelectableValue } from '@grafana/data';
 import { TwinMakerDataSource } from '../datasource';
 import { defaultQuery, TwinMakerDataSourceOptions } from '../types';
@@ -28,6 +39,7 @@ import {
 import { getTemplateSrv } from '@grafana/runtime';
 import { getVariableOptions } from 'common/variables';
 import FilterQueryEditor from './FilterQueryEditor';
+import { BlurTextInput } from './BlurTextInput';
 
 export const firstLabelWidth = 18;
 
@@ -39,6 +51,8 @@ interface State {
   entity?: SelectableComponentInfo[];
   entityLoading?: boolean;
   topics?: TwinMakerPanelTopicInfo[];
+  invalidInterval?: boolean;
+  hasStreaming?: boolean;
 }
 
 export class QueryEditor extends PureComponent<Props, State> {
@@ -47,14 +61,16 @@ export class QueryEditor extends PureComponent<Props, State> {
   constructor(props: Props) {
     super(props);
     this.panels = getTwinMakerDashboardManager().listTwinMakerPanels();
-    this.state = {};
+    this.state = {
+      invalidInterval: false,
+    };
   }
 
   componentDidMount() {
     this.loadWorkspaceInfo();
     this.loadEntityInfo(this.props.query);
     this.loadTopicInfo(this.props.query);
-    this.setState({ templateVars: getVariableOptions() });
+    this.setState({ templateVars: getVariableOptions({ keepVarSyntax: true }) });
   }
 
   loadWorkspaceInfo = async () => {
@@ -241,6 +257,26 @@ export class QueryEditor extends PureComponent<Props, State> {
     onRunQuery();
   };
 
+  onToggleStream = () => {
+    const { onChange, query, onRunQuery } = this.props;
+    onChange({ ...query, isStreaming: !query.isStreaming });
+    onRunQuery();
+  };
+
+  onIntervalChange = (value?: string) => {
+    const { onChange, query, onRunQuery } = this.props;
+    // not sending input less than 5 secs
+    if (value?.length && +value < 5) {
+      this.setState({ invalidInterval: true });
+    } else {
+      this.setState({ invalidInterval: false });
+    }
+    onChange({
+      ...query,
+      intervalStreaming: value,
+    });
+    onRunQuery();
+  };
   renderEntitySelector(query: TwinMakerQuery, isClearable: boolean) {
     const entity = getSelectionInfo(query.entityId, this.state.workspace?.entities, this.state.templateVars);
     return (
@@ -292,6 +328,7 @@ export class QueryEditor extends PureComponent<Props, State> {
             placeholder={placeholder}
           />
         </InlineField>
+        {this.renderStreamingInputs(query)}
       </InlineFieldRow>
     );
   }
@@ -312,6 +349,7 @@ export class QueryEditor extends PureComponent<Props, State> {
             formatCreateLabel={(v) => `Component Name: ${v}`}
           />
         </InlineField>
+        {this.renderStreamingInputs(query)}
       </InlineFieldRow>
     );
   }
@@ -348,7 +386,41 @@ export class QueryEditor extends PureComponent<Props, State> {
             isClearable={isClearable}
           />
         </InlineField>
+        {this.renderStreamingInputs(query)}
       </InlineFieldRow>
+    );
+  }
+
+  renderStreamingInputs(query: TwinMakerQuery) {
+    if (!this.props.datasource.grafanaLiveEnabled) {
+      return null;
+    }
+
+    return (
+      <>
+        <InlineField label="Stream" tooltip="Polling data in an interval">
+          <InlineSwitch value={Boolean(query.isStreaming)} onChange={this.onToggleStream} />
+        </InlineField>
+        <InlineField
+          label="Interval"
+          tooltip="Set an interval in seconds to stream data, min 5s, default 30s"
+          disabled={!query.isStreaming}
+          invalid={this.state.invalidInterval}
+        >
+          <>
+            <BlurTextInput
+              width={8}
+              placeholder="30"
+              value={query.intervalStreaming ?? ''}
+              onChange={this.onIntervalChange}
+              numeric={true}
+            />
+            {this.state.invalidInterval && (
+              <FieldValidationMessage>Interval must be at least 5s</FieldValidationMessage>
+            )}
+          </>
+        </InlineField>
+      </>
     );
   }
 
@@ -363,10 +435,11 @@ export class QueryEditor extends PureComponent<Props, State> {
         >
           <Input
             className="width-15"
-            value={query.maxResults}
+            value={query.maxResults && query.maxResults > 0 ? query.maxResults : ''}
             type="number"
             onChange={this.onMaxResultsChange}
             placeholder="50"
+            min="1"
           />
         </InlineField>
       </InlineFieldRow>