+++ src/panels/layout/LayoutPanel.tsx
@@ -18,7 +18,6 @@ import { ReplaySubject, of, throwError } from 'rxjs';
 import { getCurrentDashboard } from 'common/dashboard';
 import { doQueryUpdate, mergeDashboard } from './merge';
 import { TWINMAKER_PANEL_TYPE_ID } from 'common/constants';
-import { CustomScrollbar } from '@grafana/ui';
 
 const twinMakerQueries: Array<Partial<TwinMakerQuery>> = [
   { queryType: TwinMakerQueryType.ListWorkspace },
@@ -46,7 +45,7 @@ export class LayoutPanel extends Component<Props, State> {
   ds?: TwinMakerDataSource;
   loading?: boolean;
   initalized?: boolean;
-  dashboardUid: string;
+  dashboardUid: string | undefined;
   error?: string;
 
   constructor(props: Props) {
@@ -61,14 +60,16 @@ export class LayoutPanel extends Component<Props, State> {
     };
 
     const dashboard = getCurrentDashboard();
-    this.dashboardUid = dashboard.uid;
-    for (const panel of dashboard.panels) {
-      if (panel.id !== this.props.id && panel.type === TWINMAKER_PANEL_TYPE_ID.LAYOUT) {
-        const isEditor = window.location.href.includes('editPanel');
-        if (!isEditor) {
-          this.error = 'Only one layout panel allowed in a dashboard';
+    if (dashboard) {
+      this.dashboardUid = dashboard.uid;
+      for (const panel of dashboard.panels) {
+        if (panel.id !== this.props.id && panel.type === TWINMAKER_PANEL_TYPE_ID.LAYOUT) {
+          const isEditor = window.location.href.includes('editPanel');
+          if (!isEditor) {
+            this.error = 'Only one layout panel allowed in a dashboard';
+          }
+          break;
         }
-        break;
       }
     }
 
@@ -228,22 +229,20 @@ export class LayoutPanel extends Component<Props, State> {
     switch (options.show) {
       case DisplayMode.EntityDetails:
         return (
-          <CustomScrollbar autoHeightMin="100%">
-            <div>
-              <table>
-                <tr>
-                  <th>EntityName: &nbsp;</th>
-                  <td>{entityInfo?.EntityName}</td>
-                </tr>
-                <tr>
-                  <th>ComponentTypeId: &nbsp;</th>
-                  <td>{componentInfo?.ComponentTypeId}</td>
-                </tr>
-              </table>
-              {componentInfo && <pre>{JSON.stringify(componentInfo, null, 2)}</pre>}
-              {entityInfo && !componentInfo && <pre>{JSON.stringify(entityInfo, null, 2)}</pre>}
-            </div>
-          </CustomScrollbar>
+          <div>
+            <table>
+              <tr>
+                <th>EntityName: &nbsp;</th>
+                <td>{entityInfo?.EntityName}</td>
+              </tr>
+              <tr>
+                <th>ComponentTypeId: &nbsp;</th>
+                <td>{componentInfo?.ComponentTypeId}</td>
+              </tr>
+            </table>
+            {componentInfo && <pre>{JSON.stringify(componentInfo, null, 2)}</pre>}
+            {entityInfo && !componentInfo && <pre>{JSON.stringify(entityInfo, null, 2)}</pre>}
+          </div>
         );
 
       case DisplayMode.ComponentTypeId: