+++ CHANGELOG.md
@@ -1,8 +1,22 @@
 # Changelog
 
+## v1.2.1
+
+- Fix for paginaton issue that lead to partial data on the dashboard
+- Fix minor bugs with the Scene Viewer and Video Player panels
+
+## v1.2.0
+
+- Live time series history streaming
+- Set max number of alarms to retrieve data for in the Get Alarms query
+- Set propertyName using a template variable
+- Avoid unnecessary panel refreshes on the dashboard
+- Show Motion Indicator in the scene viewer
+- Fixed Model Shader functionality in the scene viewer
+
 ## v1.1.2
 
-- Increase the size and brightness of the tag selection graphic 
+- Increase the size and brightness of the tag selection graphic
 - Color the outer circle to be blue for tags
 - Make video tag consistent with other tags
 - Black icon issue for tags in FireFox